Commit Update: SOBJ_APP:Adding option to save fake pem file
There was a requirement for generating a fake .pem file
for the key generated when integrating with kubernetes.
Added a command line option "-w" to create a fake .pem file.

./sobj_app -G -m <mechanism-ID> -s <key-size> -l <key-label> -i <key-ID> [-w <fake-pem-file>]

Now this is saving key-ID in the private part of RSA key.
Which will be decoded while doing RSA private operations in Engine.

Signed-off-by: Sahil Malhotra <sahil.malhotra@nxp.com>